# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MeshCore network observation backend. Ingests LoRa packets from MQTT brokers, stores in PostgreSQL, and streams live events via WebSocket.",
        "title": "MeshCore Beacon API",
        "termsOfService": "https://github.com/MeshCore-Beacon/beacon-server",
        "contact": {
            "name": "MeshCore Beacon",
            "url": "https://github.com/MeshCore-Beacon/beacon-server"
        },
        "license": {
            "name": "AGPL-3-or-later"
        },
        "version": "1.4.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/packets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packets"
                ],
                "summary": "List packets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by payload type integer",
                        "name": "payloadType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by payload type name (advert, grp_txt, txt_msg, trace, anon_req)",
                        "name": "payloadTypeName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by route type (0=transport_flood, 1=flood, 2=direct, 3=transport_direct)",
                        "name": "routeType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by single IATA code (case-insensitive)",
                        "name": "iata",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by multiple IATA codes, comma-separated e.g. YVR,YYJ",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport scope name e.g. %23bc (URL-encoded #bc)",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID, expands to member IATAs",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug, expands to member IATAs",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by first_heard_at \u003e= since (epoch ms)",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by first_heard_at \u003c= until (epoch ms)",
                        "name": "until",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "last_heard_at epoch ms of last item for pagination",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/packets/backfill": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packets"
                ],
                "summary": "Backfill packets after a given observation ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Return packets with observations after this ID (use last WS event observation ID)",
                        "name": "afterObservationId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Filter by payload type integer",
                        "name": "payloadType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by payload type name",
                        "name": "payloadTypeName",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by route type",
                        "name": "routeType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by IATA code(s), comma-separated",
                        "name": "iatas",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by region slug",
                        "name": "region",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by region ID",
                        "name": "regionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by transport scope name",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketSummary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
        "/packets/{packetHash}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Packets"
                ],
                "summary": "Get full packet detail. For trace packets (payloadType=9), includes resolvedRoute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Packet hash (hex)",
                        "name": "packetHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketDoc": {
            "type": "object",
            "properties": {
                "channelHash": {
                    "type": "string"
                },
                "decrypted": {
                    "type": "boolean"
                },
                "firstHeardAt": {
                    "type": "integer"
                },
                "firstToLastMs": {
                    "type": "integer"
                },
                "header": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketHeader"
                },
                "lastHeardAt": {
                    "type": "integer"
                },
                "observationCount": {
                    "type": "integer"
                },
                "observations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketObservationDetail"
                    }
                },
                "originPubkey": {
                    "type": "string"
                },
                "packetHash": {
                    "type": "string"
                },
                "parsedPayload": {
                    "type": "object",
                    "additionalProperties": true
                },
                "rawPayload": {
                    "type": "string"
                },
                "resolvedRoute": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop"
                    }
                },
                "scope": {
                    "type": "string"
                },
                "transportCodes": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketTransportCodes"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketHeader": {
            "type": "object",
            "properties": {
                "payloadType": {
                    "description": "bits 2-5",
                    "type": "integer"
                },
                "payloadTypeName": {
                    "description": "advert, request, group_text, etc.",
                    "type": "string"
                },
                "payloadVersion": {
                    "description": "bits 6-7",
                    "type": "integer"
                },
                "raw": {
                    "description": "hex-encoded single byte",
                    "type": "string"
                },
                "routeType": {
                    "description": "bits 0-1",
                    "type": "integer"
                },
                "routeTypeName": {
                    "description": "FLOOD, DIRECT, TRANSPORT_FLOOD, TRANSPORT_DIRECT",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketLatestObserver": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketObservationDetail": {
            "type": "object",
            "properties": {
                "heardAt": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "iata": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "observerId": {
                    "type": "string"
                },
                "observerName": {
                    "type": "string"
                },
                "pathBytes": {
                    "description": "hex-encoded accumulated path hashes",
                    "type": "string"
                },
                "pathLength": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketPathLength"
                },
                "propagationTimeMs": {
                    "description": "ms since first observation; 0 for first",
                    "type": "integer"
                },
                "radio": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketRadio"
                },
                "resolvedPath": {
                    "description": "per-observation resolved path hashes",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop"
                    }
                },
                "rssi": {
                    "type": "integer"
                },
                "snr": {
                    "type": "number"
                },
                "sourceBroker": {
                    "type": "string"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketPathLength": {
            "type": "object",
            "properties": {
                "hashSize": {
                    "description": "per-hop hash size in bytes (1, 2, or 3)",
                    "type": "integer"
                },
                "hopCount": {
                    "description": "number of path hashes present",
                    "type": "integer"
                },
                "raw": {
                    "description": "hex-encoded single byte",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketRadio": {
            "type": "object",
            "properties": {
                "bandwidthKhz": {
                    "type": "number"
                },
                "codingRate": {
                    "type": "integer"
                },
                "freqMhz": {
                    "type": "number"
                },
                "spreadFactor": {
                    "type": "integer"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketSummary": {
            "type": "object",
            "properties": {
                "firstHeardAt": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "lastHeardAt": {
                    "description": "epoch ms",
                    "type": "integer"
                },
                "latestObserver": {
                    "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.PacketLatestObserver"
                },
                "observationCount": {
                    "type": "integer"
                },
                "packetHash": {
                    "description": "hex-encoded",
                    "type": "string"
                },
                "payloadType": {
                    "type": "integer"
                },
                "payloadTypeName": {
                    "type": "string"
                },
                "routeType": {
                    "type": "integer"
                },
                "routeTypeName": {
                    "type": "string"
                },
                "scope": {
                    "description": "matched transport scope name e.g. \"#bc\"",
                    "type": "string"
                },
                "summary": {
                    "description": "human-readable payload summary",
                    "type": "string"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.PacketTransportCodes": {
            "type": "object",
            "properties": {
                "regionCode": {
                    "type": "integer"
                },
                "subRegionCode": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedHop": {
            "type": "object",
            "properties": {
                "confidence": {
                    "description": "\"high\", \"ambiguous\", or \"none\"",
                    "type": "string"
                },
                "nodes": {
                    "description": "empty for \"none\", one for \"high\", multiple for \"ambiguous\"",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode"
                    }
                },
                "snr": {
                    "type": "number"
                }
            }
        },
        "github_com_MeshCore-Beacon_beacon-server_internal_api.ResolvedNode": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "publicKey": {
                    "description": "hex-encoded prefix used for resolution",
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "e.g. \"not_found\", \"bad_request\"",
                    "type": "string"
                },
                "message": {
                    "description": "e.g. \"channel not found\"",
                    "type": "string"
                }
            }
        },
...
    },
...
}